Fix the build by marking dependencies as compile
Thanks to the answer to my question at http://stackoverflow.com/q/21919276/578288 . I tried changing `runtime` to `compile` before, but it didn’t help, so I thought that `runtime` was correct. It turned out that I needed to `./gradlew clean` to be extra sure.

This commit includes adding the Java file I tested building with